add advanced period support
The existing period support requires that the base value be an ordinal.
To handle more advanced cases we can allow the base value to be a
specifier and then pull out some existing specifier variants into a root
specifier. Namely, a root specifier can contain a specifier, a period,
or a named point. A specifier can be all, point, range, or named range.

This adds support for cases such as '*/2', '10-20/2', and 'Mon-Thurs/2'.

Signed-off-by: Grant Elbert <grant.elbert@smartthings.com>